artificial intelligence problems
There are two artificial intelligence problems in this code.
A human trying to get out of a labirynth.
And the famous Hanoi towers.
https://en.wikipedia.org/wiki/Tower_of_Hanoi
It is implemented in python.
A star (A*) algorithm is used to solve the above problems.
https://en.wikipedia.org/wiki/A*_search_algorithm
For heuristics i used, manhattan distance and euclidean distance to
compare the results.